Refactor useBeers to fetch express API
The beerRoutes file now has a get request that will request all
the beers that are held within the database in the beers collection.
On the front, the useBeers hook was refactored to fetch from the
express api instead of from the third-party api. The Shop page is
now rendering out all the beers, but from the database instead.