Start functionality for adding beer to cart
On the front, on the Shop page, an event handler was attached to
each button, eventually allowing for a POST request to be made. On
the back, a GET request was made for the cartRoutes, and the start
of a POST request was begun, but still needs a lot of work.