Create new addedBeer model
The new addedBeer model will replace the Cart model, since it
was deemed to unnecessary to nest an array within the Cart model.
Now the addedBeer model will handle all the added beers, and now
will keep track of the quantity. On the front end, on the Shop and
Cart pages, there was refactoring done to make the new
addedBeerRoutes integrate properly.